requirements.txt
langgraph==0.4.7
langchain-core>=0.1.50
langchain-openai>=0.1.6
openai>=1.30.0
pydantic>=2.0
reportlab>=4.0
sendgrid>=6.11.0
nest_asyncio>=1.6
Flask>=3.0
pyppeteer>=1.0
pygraphviz>=1.12
graphviz>=0.20.1
python-dotenv>=1.0.0

